Add optional JSON output support
Enable JSON output for efibootmgr. This provides an interface for third
party tools so they do not need to scrap the existing textual output.

This feature is optional and is enabled by using JSON=1 when compiling:

	make ... JSON=1

Add parameter '-j'/--json' to use JSON output. If efibootmgr is not
built with JSON output support it will print the following to stderr:

	"JSON support is not built-in"

This feature adds a new optional dependency with libjansson.

Signed-off-by: Jose M. Guisado <jguisado@soleta.eu>
Co-authored-by: Alejandro Sirgo Rica <asirgo@soleta.eu>